pyproject.toml
[project]
name = "memescope"
version = "0.1.0"
description = ""
authors = [
    {name = "Granth Agarwal",email = "heygranth@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "anyio (==4.9.0)",
    "asgiref (==3.8.1)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.3)",
    "certifi (==2025.4.26)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.2)",
    "cleo (==2.1.0)",
    "crashtest (==0.4.1)",
    "cryptography (==44.0.3)",
    "distlib (==0.3.9)",
    "django (==5.2.1)",
    "djangorestframework (==3.16.0)",
    "dulwich (==0.22.8)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "jeepney (==0.9.0)",
    "keyring (==25.6.0)",
    "more-itertools (==10.7.0)",
    "msgpack (==1.1.0)",
    "packaging (==25.0)",
    "pbs-installer (==2025.4.9)",
    "pip (==25.1.1)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.8)",
    "poetry (==2.1.3)",
    "poetry-core (==2.1.3)",
    "praw (==7.8.1)",
    "prawcore (==2.4.0)",
    "pycparser (==2.22)",
    "pyproject-hooks (==1.2.0)",
    "rapidfuzz (==3.13.0)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "secretstorage (==3.3.3)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "sqlparse (==0.5.3)",
    "tomlkit (==0.13.2)",
    "trove-classifiers (==2025.5.9.12)",
    "typing-extensions (==4.13.2)",
    "update-checker (==0.18.0)",
    "urllib3 (==2.4.0)",
    "uv (==0.7.3)",
    "virtualenv (==20.31.2)",
    "websocket-client (==1.8.0)",
    "zstandard (==0.23.0)",
    "python-decouple (==3.8)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
